refactor: organize tree related files, render children not footer
